Include Target String in String FunctionBuilder dialogs
